add nl_hashing
separate hashing related functions in separate header
